decorator with params
